Make global input bar
